Fix broken Unicode properly this time
If the strings are identical, there must be something else that differs.

Turns out, it's the channel name ($chl) – in some invocations, it has
the is_utf8 flag set (e.g. when idle-chatting).

When POE::Component::IRC creates the final line with "$state $to :$msg",
Perl sets is_utf8 on it if /any/ of the input strings had it, but won't
utf8-encode strings that didn't. (Instead, they get included as if each
byte was an individual codepoint.)

To fix this, always make sure POE receives only raw byte strings, by
encoding them when necessary.