Fix behavior after adding new large var values
Perl has an idiosyncrasy when it comes to foreach loops: If you try to
foreach over a hash element that doesn't exist, Perl will simply create
it for you. Since Bucket relies on checking whether a given variable's
hash contains a "vals" array, Perl automatically creating that array did
bad things to Bucket's logic, tricking Bucket into thinking it was adding
the value into a short var instead and forcing all references to that var
to bypass the large var cache in favor of the incomplete list in memory.
Well, at least until Bucket's next nap (restart).

Resolves #62, where you may find a more detailed discussion and the saga
of how this bug was uncovered.