Item deletion message: destroy -> delete
Change the confirmation for item deletion from "destroying $item" to "deleting item $item", which is more consistent with the command used to do so ("delete item #20"). The last few times I've wanted to delete an item it took me three tries to get there ("destroy item #20", "destroy #20") and I feel this might help.